+++ src/WebCompiler/Compile/LessOptions.cs
@@ -18,6 +18,10 @@ namespace WebCompiler
         /// </summary>
         protected override void LoadSettings(Config config)
         {
+            var autoPrefix = GetValue(config, "autoPrefix");
+            if (autoPrefix != null)
+                AutoPrefix = autoPrefix;
+
             var strictMath = GetValue(config, "strictMath");
             if (strictMath != null)
                 StrictMath = strictMath.ToLowerInvariant() == trueStr;
@@ -27,8 +31,8 @@ namespace WebCompiler
                 StrictUnits = strictUnits.ToLowerInvariant() == trueStr;
 
             var rootPath = GetValue(config, "rootPath");
-            if (RootPath != null)
-                RootPath = RootPath;
+            if (rootPath != null)
+                RootPath = rootPath;
 
             var relativeUrls = GetValue(config, "relativeUrls");
             if (relativeUrls != null)
@@ -44,6 +48,13 @@ namespace WebCompiler
         }
 
         /// <summary>
+        /// Autoprefixer will use the data based on current browser popularity and
+        /// property support to apply prefixes for you.
+        /// </summary>
+        [JsonProperty("autoPrefix")]
+        public string AutoPrefix { get; set; } = "";
+
+        /// <summary>
         /// Without this option on Less will try and process all maths in your CSS.
         /// </summary>
         [JsonProperty("strictMath")]
@@ -56,14 +67,14 @@ namespace WebCompiler
         public bool StrictUnits { get; set; } = false;
 
         /// <summary>
-        /// This option allows you to re-write URL's in imported files so that the URL is always 
+        /// This option allows you to re-write URL's in imported files so that the URL is always
         /// relative to the base imported file.
         /// </summary>
         [JsonProperty("relativeUrls")]
         public bool RelativeUrls { get; set; } = true;
 
         /// <summary>
-        /// Allows you to add a path to every generated import and url in your css. 
+        /// Allows you to add a path to every generated import and url in your css.
         /// This does not affect less import statements that are processed,
         /// just ones that are left in the output css.
         /// </summary>